+++ scm_helper/api.py
@@ -1,5 +1,6 @@
 """Interface to SCM API."""
 import os.path
+import platform
 from datetime import date, datetime
 from pathlib import Path
 from shutil import copyfile
@@ -31,14 +32,15 @@ from scm_helper.config import (
     URL_MEMBERS,
     URL_ROLES,
     URL_SESSIONS,
+    URL_WHO,
     USER_AGENT,
+    WHO,
     get_config,
     verify_schema,
     verify_schema_data,
 )
-from scm_helper.crypto import Crypto
 from scm_helper.default import create_default_config
-from scm_helper.entity import Entities
+from scm_helper.entity import Entities, Who
 from scm_helper.groups import Groups
 from scm_helper.issue import debug, set_debug_level
 from scm_helper.lists import Lists
@@ -53,6 +55,7 @@ class API:
     """Main SCM object."""
 
     # pylint: disable=too-many-instance-attributes
+    # pylint: disable=too-many-public-methods
     # Need them all!
 
     def __init__(self, issues):
@@ -74,6 +77,7 @@ class API:
         self.issue_handler = issues
         self.fixable = []
         self.crypto = None
+        self.ipad = False
 
         self.today = datetime.now()
         q_month = (int((self.today.month - 1) / 3) * 3) + 1
@@ -81,6 +85,9 @@ class API:
         self.eoy = datetime(int(q_year), 12, 31)
         offset = datetime(int(q_year), int(q_month), 1)
         self.q_offset = (self.today - offset).days
+
+        if "iPad" in platform.machine():
+            self.ipad = True
 
     def get_config_file(self):
         """Read configuration file."""
@@ -100,7 +107,7 @@ class API:
         except EnvironmentError:
             notify(f"Cannot open configuration file: {cfg}\n")
             return False
-        except yaml.scanner.ScannerError as error:
+        except (yaml.scanner.ScannerError, yaml.parser.ParserError) as error:
             notify(f"Error in configuration file: {error}\n")
             return False
 
@@ -111,11 +118,19 @@ class API:
 
     def get_config(self, password):
         """Get API key."""
+        # pylint: disable=import-outside-toplevel
         if self._config is None:
             if self.get_config_file() is False:
                 return False
 
-        self.crypto = Crypto(self._config[C_CLUB], password)  # Salt
+        if self.ipad:
+            from scm_helper.ipad import Crypto
+
+            self.crypto = Crypto(self._config[C_CLUB], password)  # Salt
+        else:
+            from scm_helper.crypto import Crypto
+
+            self.crypto = Crypto(self._config[C_CLUB], password)  # Salt
 
         home = str(Path.home())
 
@@ -133,6 +148,10 @@ class API:
 
     def initialise(self, password):
         """Initialise."""
+
+        if self.ipad:
+            password = "dummy"  # Can't to crypto on iPad
+
         if self.get_config(password) is False:
             return False
 
@@ -176,6 +195,12 @@ class API:
             name = name.lower()
             self.class_byname[name] = entity
 
+        # Finally who's who
+        entity = Who(self, WHO, URL_WHO)
+        self.backup_classes.append(entity)
+        name = WHO.lower()
+        self.class_byname[name] = entity
+
         return True
 
     def get_data(self, backup):
@@ -196,6 +221,10 @@ class API:
     def get_members_only(self):
         """Get member data."""
         self.members.get_data()
+
+    def se_check(self):
+        """Get member data."""
+        return self.members.se_check()
 
     def linkage(self):
         """Set up cross reference links between Entities."""
@@ -225,6 +254,10 @@ class API:
 
     def restore(self, xclass):
         """Restore data..."""
+        if self.ipad:
+            notify("Not implemented on iPad")
+            return False
+
         xclass = xclass.lower()
         if xclass in self.class_byname:
             item = self.class_byname[xclass]
@@ -235,6 +268,10 @@ class API:
 
     def dump(self, xclass):
         """Dump data..."""
+        if self.ipad:
+            notify("Not implemented on iPad")
+            return False
+
         f_csv = "CSV"
         f_json = "JSON"
         xlist = [f_json, f_csv]
@@ -274,12 +311,16 @@ class API:
 
     def backup_data(self):
         """Backup."""
+        if self.ipad:
+            notify("Not implemented on iPad")
+            return False
+
         if self.get_data(True) is False:
             return False
 
         backup = self.classes + self.backup_classes
         for aclass in backup:
-            if self.crypto.encrypt_file(aclass.name, aclass.json) is False:
+            if self.crypto.encrypt_backup(aclass.name, aclass.json) is False:
                 return False
 
         # Backup config file too.
@@ -301,10 +342,14 @@ class API:
 
     def decrypt(self, xdate):
         """Decrypt file."""
+        if self.ipad:
+            notify("Not implemented on iPad")
+            return False
+
         restore = self.classes + self.backup_classes
 
         for aclass in restore:
-            decrypted = self.crypto.decrypt_file(aclass.name, xdate)
+            decrypted = self.crypto.decrypt_backup(aclass.name, xdate)
             if decrypted is None:
                 return False
             aclass.parse_data(decrypted)
@@ -332,7 +377,8 @@ class API:
 
         length = len(self.fixable)
         if length > 0:
-            output += f"\n{length} fixable errors."
+            output += f"\n{length} fixable errors...\n"
+            output += self.list_fixes()
 
         output += "\n"
         return output
@@ -420,6 +466,18 @@ class API:
         self.fixable = []
         return True
 
+    def list_fixes(self):
+        """List any fixes."""
+        if len(self.fixable) == 0:
+            notify("Nothing to fix\n")
+            return False
+
+        res = ""
+        for fix in self.fixable:
+            res += f"{fix.name}: {fix.fixmsg}\n"
+
+        return res
+        
     def option(self, option):
         """Options."""
         if option in self._options: