+++ scm_helper/notify.py
@@ -1,6 +1,11 @@
 """Notify."""
 import sys
-from tkinter import END, messagebox, simpledialog
+
+try:
+    from tkinter import END, messagebox, simpledialog
+except ImportError:
+    pass  # ignore it, we'll be running command line only
+
 
 WHERE = None